# Performance Optimization Suite Phase 2: Achieve 98% search time reduction (53s → 0.9s)