--- a/src/language/tree_cache.rs
+++ b/src/language/tree_cache.rs
@@ -1,7 +1,11 @@
 use anyhow::{Context, Result};
-use std::collections::HashMap;
+use lru::LruCache;
+use std::num::NonZeroUsize;
 use std::sync::Mutex;
 use tree_sitter::Tree;
+
+// PHASE 4 OPTIMIZATION: Use environment variable or CPU count for cache size
+const DEFAULT_CACHE_SIZE: usize = 2000;
 
 lazy_static::lazy_static! {
     /// A cache for parsed syntax trees to avoid redundant parsing
@@ -9,7 +13,14 @@ lazy_static::lazy_static! {
     /// This cache stores parsed ASTs keyed by file path and content hash.
     /// When the same file is parsed multiple times, this avoids the overhead
     /// of re-parsing unchanged files.
-    static ref TREE_CACHE: Mutex<HashMap<String, (Tree, u64)>> = Mutex::new(HashMap::new());
+    /// PHASE 4 OPTIMIZATION: Use LRU cache with bounded size to prevent memory bloat
+    static ref TREE_CACHE: Mutex<LruCache<String, (Tree, u64)>> = {
+        let cache_size = std::env::var("PROBE_TREE_CACHE_SIZE")
+            .ok()
+            .and_then(|s| s.parse().ok())
+            .unwrap_or(DEFAULT_CACHE_SIZE);
+        Mutex::new(LruCache::new(NonZeroUsize::new(cache_size).unwrap()))
+    };
 
     /// A counter for cache hits, used for testing
     static ref CACHE_HITS: Mutex<usize> = Mutex::new(0);
@@ -66,7 +77,8 @@ pub fn get_or_parse_tree(
         let mut cache = TREE_CACHE
             .lock()
             .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if let Some((cached_tree, cached_hash)) = cache.get(file_path) {
+        // PHASE 4 OPTIMIZATION: Use peek to check without updating LRU order first
+        if let Some((_cached_tree, cached_hash)) = cache.peek(file_path) {
             if cached_hash == &content_hash {
                 // Increment cache hit counter
                 {
@@ -79,10 +91,13 @@ pub fn get_or_parse_tree(
                 if debug_mode {
                     println!("[DEBUG] Cache hit for file: {file_path}");
                 }
+                // PHASE 4 OPTIMIZATION: Now update LRU order with get
+                let (cached_tree, _) = cache.get(file_path).unwrap();
                 return Ok(cached_tree.clone());
             } else {
                 // Content changed, explicitly remove the old entry
-                cache.remove(file_path);
+                // PHASE 4 OPTIMIZATION: LRU cache uses pop instead of remove
+                cache.pop(file_path);
                 if debug_mode {
                     println!("[DEBUG] Cache invalidated for file: {file_path} (content changed)");
                 }
@@ -102,7 +117,8 @@ pub fn get_or_parse_tree(
         let mut cache = TREE_CACHE
             .lock()
             .unwrap_or_else(|poisoned| poisoned.into_inner());
-        cache.insert(file_path.to_string(), (tree.clone(), content_hash));
+        // PHASE 4 OPTIMIZATION: LRU cache automatically evicts old entries
+        cache.put(file_path.to_string(), (tree.clone(), content_hash));
 
         if debug_mode {
             println!("[DEBUG] Cached parsed tree for file: {file_path}");
@@ -142,7 +158,8 @@ pub fn get_or_parse_tree_pooled(file_path: &str, content: &str, extension: &str)
         let mut cache = TREE_CACHE
             .lock()
             .unwrap_or_else(|poisoned| poisoned.into_inner());
-        if let Some((cached_tree, cached_hash)) = cache.get(file_path) {
+        // PHASE 4 OPTIMIZATION: Use peek to check without updating LRU order first
+        if let Some((_cached_tree, cached_hash)) = cache.peek(file_path) {
             if cached_hash == &content_hash {
                 // Increment cache hit counter
                 {
@@ -155,10 +172,13 @@ pub fn get_or_parse_tree_pooled(file_path: &str, content: &str, extension: &str)
                 if debug_mode {
                     println!("[DEBUG] Tree cache hit for file: {file_path}");
                 }
+                // PHASE 4 OPTIMIZATION: Now update LRU order with get
+                let (cached_tree, _) = cache.get(file_path).unwrap();
                 return Ok(cached_tree.clone());
             } else {
                 // Content changed, explicitly remove the old entry
-                cache.remove(file_path);
+                // PHASE 4 OPTIMIZATION: LRU cache uses pop instead of remove
+                cache.pop(file_path);
                 if debug_mode {
                     println!(
                         "[DEBUG] Tree cache invalidated for file: {file_path} (content changed)"
@@ -187,7 +207,8 @@ pub fn get_or_parse_tree_pooled(file_path: &str, content: &str, extension: &str)
         let mut cache = TREE_CACHE
             .lock()
             .unwrap_or_else(|poisoned| poisoned.into_inner());
-        cache.insert(file_path.to_string(), (tree.clone(), content_hash));
+        // PHASE 4 OPTIMIZATION: LRU cache automatically evicts old entries
+        cache.put(file_path.to_string(), (tree.clone(), content_hash));
 
         if debug_mode {
             println!("[DEBUG] Cached parsed tree for file: {file_path}");
@@ -233,7 +254,7 @@ pub fn invalidate_cache_entry(file_path: &str) {
         .unwrap_or_else(|poisoned| poisoned.into_inner());
     let debug_mode = std::env::var("DEBUG").unwrap_or_default() == "1";
 
-    if cache.remove(file_path).is_some() && debug_mode {
+    if cache.pop(file_path).is_some() && debug_mode {
         println!("[DEBUG] Removed file from cache: {file_path}");
     }
 }
@@ -264,7 +285,7 @@ pub fn is_in_cache(file_path: &str) -> bool {
     let cache = TREE_CACHE
         .lock()
         .unwrap_or_else(|poisoned| poisoned.into_inner());
-    cache.contains_key(file_path)
+    cache.contains(file_path)
 }
 
 /// Reset the cache hit counter (for testing)