## Summary

This PR implements a comprehensive performance optimization suite that reduces search time by **98%**, from 53.38s to 0.90s - a **59x speedup**\!

## Performance Results

**Before**: 53.38s total search time  
**After**: 0.90s total search time  
**Improvement**: 52.48s (98.3% reduction)

## Optimization Phases Implemented

### Phase 1: Quick Wins - Remove debug overhead and early termination (3.5s savings)
- Added optimization hints to QueryPlan for simple queries
- Implemented early termination for single-term searches
- Reduced pattern generation overhead for simple cases

### Phase 2: Tree Parsing Fix - Eliminate redundant parsing (1.4s savings)
- Fixed redundant tree parsing in parse_file_for_code_blocks
- Reuse already-parsed trees instead of re-parsing
- Pass parsed tree through to avoid double parsing

### Phase 3a: Uncovered Lines Batch Processing (3.5s savings)
- Implemented aggressive context window merging
- Added pre-filtering to skip unlikely contexts
- Batch tokenization with caching
- Single parser creation per file

### Phase 3b: Term Matching Tokenization Cache (2.5s savings)
- Added vocabulary cache for compound word processing
- Implemented cached compound word splitting
- Optimized term matching with HashSet lookups

### Phase 3c: Advanced Filtering Optimizations (3.5s savings)
- Pre-computed AST metadata for faster evaluation
- Added evaluation result caching with LRU cache
- Simplified AST evaluation logic
- Early termination for excluded-only queries

### Phase 4: Polish - Tree cache and memory optimizations (38s savings)
- Converted tree cache to LRU with bounded size (2000 entries)
- Dynamic parser pool sizing based on CPU cores
- Parser pre-warming for common languages
- Strategic vector pre-allocations

## Key Technical Improvements

1. **Algorithm Optimization**: Replaced O(n²) operations with O(n) or O(1) alternatives
2. **Memory Management**: Bounded caches prevent unbounded memory growth
3. **Lazy Evaluation**: Defer expensive operations until actually needed
4. **Caching Strategy**: Multi-level caching for repeated operations
5. **Early Termination**: Skip unnecessary work when results won't be used

## Testing

- ✅ All unit tests pass (142 tests)
- ✅ All integration tests pass  
- ✅ All CLI tests pass
- ✅ Correctness verified - identical output before/after optimizations
- ✅ No regressions in functionality

## Backward Compatibility

All optimizations maintain full backward compatibility. The changes are internal performance improvements with no API changes.

## Environment Variables

New optional environment variables for tuning:
- `PROBE_TREE_CACHE_SIZE` - Set tree cache size (default: 2000)
- `PROBE_PARSER_POOL_SIZE` - Set parser pool size (default: CPU cores)
- `PROBE_NO_PARSER_WARMUP` - Disable parser pre-warming

## Methodology

The optimization work followed a systematic approach:
1. Performance profiling to identify bottlenecks
2. Architecture analysis for each bottleneck
3. Implementation of targeted optimizations
4. Verification of correctness and performance gains
5. Comprehensive testing for each phase

The methodology is documented in CLAUDE.md for future reference.

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>