@buger 2025-07-31
/engineer pls fix latest clippy warnings:

    Checking xml-schema v0.2.0
error: this `map_or` can be simplified
   --> src\language\parser_pool.rs:219:31
    |
219 |             let checked_out = stats.get("rs").map_or(true, |&count| count == 0);
    |                               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_map_or
    = note: `-D clippy::unnecessary-map-or` implied by `-D warnings`
    = help: to override `-D warnings` add `#[allow(clippy::unnecessary_map_or)]`
help: use is_none_or instead
    |
219 -             let checked_out = stats.get("rs").map_or(true, |&count| count == 0);
219 +             let checked_out = stats.get("rs").is_none_or(|&count| count == 0);
    |

error: could not compile `probe-code` (lib test) due to 1 previous error
warning: build failed, waiting for other jobs to finish...
Error: Process completed with exit code 1.